test(extension): extract extension-fixtures.ts
Move fixture definitions, helpers, and shared types out of
extension.spec.ts into a dedicated extension-fixtures.ts so the
spec file only contains tests.
