+++ packages/extension/tests/extension.spec.ts
@@ -15,188 +15,7 @@
  */
 
 import fs from 'fs/promises';
-import path from 'path';
-import { chromium } from 'playwright';
-import { spawn } from 'child_process';
-import { test as base, expect } from '../../playwright-mcp/tests/fixtures';
-
-import type { Client } from '@modelcontextprotocol/sdk/client/index.js';
-import type { BrowserContext } from 'playwright';
-import type { StartClient } from '../../playwright-mcp/tests/fixtures';
-
-type BrowserWithExtension = {
-  userDataDir: string;
-  launch: (mode?: 'disable-extension') => Promise<BrowserContext>;
-};
-
-type CliResult = {
-  output: string;
-  error: string;
-};
-
-type TestFixtures = {
-  browserWithExtension: BrowserWithExtension,
-  pathToExtension: string,
-  startExtensionClient: (env?: Record<string, string>) => Promise<{ browserContext: BrowserContext, client: Client }>,
-  cli: (...args: string[]) => Promise<CliResult>;
-};
-
-const extensionPublicKey = 'MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAwRsUUO4mmbCi4JpmrIoIw31iVW9+xUJRZ6nSzya17PQkaUPDxe1IpgM+vpd/xB6mJWlJSyE1Lj95c0sbomGfVY1M0zUeKbaRVcAb+/a6m59gNR+ubFlmTX0nK9/8fE2FpRB9D+4N5jyeIPQuASW/0oswI2/ijK7hH5NTRX8gWc/ROMSgUj7rKhTAgBrICt/NsStgDPsxRTPPJnhJ/ViJtM1P5KsSYswE987DPoFnpmkFpq8g1ae0eYbQfXy55ieaacC4QWyJPj3daU2kMfBQw7MXnnk0H/WDxouMOIHnd8MlQxpEMqAihj7KpuONH+MUhuj9HEQo4df6bSaIuQ0b4QIDAQAB';
-const extensionId = 'mmlmfjhmonkocbjadbfplnigmagldckm';
-
-const test = base.extend<TestFixtures>({
-  pathToExtension: async ({}, use, testInfo) => {
-    const extensionDir = testInfo.outputPath('extension');
-    const srcDir = path.resolve(__dirname, '../dist');
-    await fs.cp(srcDir, extensionDir, { recursive: true });
-    const manifestPath = path.join(extensionDir, 'manifest.json');
-    const manifest = JSON.parse(await fs.readFile(manifestPath, 'utf8'));
-    // We don't hardcode the key in manifest, but for the tests we set the key field
-    // to ensure that locally installed extension has the same id as the one published
-    // in the store.
-    manifest.key = extensionPublicKey;
-    await fs.writeFile(manifestPath, JSON.stringify(manifest, null, 2));
-    await use(extensionDir);
-  },
-
-  browserWithExtension: async ({ mcpBrowser, pathToExtension }, use, testInfo) => {
-    // The flags no longer work in Chrome since
-    // https://chromium.googlesource.com/chromium/src/+/290ed8046692651ce76088914750cb659b65fb17%5E%21/chrome/browser/extensions/extension_service.cc?pli=1#
-    test.skip('chromium' !== mcpBrowser, '--load-extension is not supported for official builds of Chromium');
-
-    let browserContext: BrowserContext | undefined;
-    const userDataDir = testInfo.outputPath('extension-user-data-dir');
-    await use({
-      userDataDir,
-      launch: async (mode?: 'disable-extension') => {
-        browserContext = await chromium.launchPersistentContext(userDataDir, {
-          channel: mcpBrowser,
-          // Opening the browser singleton only works in headed.
-          headless: false,
-          // Automation disables singleton browser process behavior, which is necessary for the extension.
-          ignoreDefaultArgs: ['--enable-automation'],
-          args: mode === 'disable-extension' ? [] : [
-            `--disable-extensions-except=${pathToExtension}`,
-            `--load-extension=${pathToExtension}`,
-          ],
-        });
-
-        // for manifest v3:
-        let [serviceWorker] = browserContext.serviceWorkers();
-        if (!serviceWorker)
-          serviceWorker = await browserContext.waitForEvent('serviceworker');
-
-        return browserContext;
-      }
-    });
-    await browserContext?.close();
-
-    // Free up disk space.
-    await fs.rm(userDataDir, { recursive: true, force: true }).catch(() => {});
-  },
-
-  startExtensionClient: async ({ browserWithExtension, startClient }, use) => {
-    await use(async (env?: Record<string, string>) => {
-      const browserContext = await browserWithExtension.launch();
-      const client = await startWithExtensionFlag(browserWithExtension, startClient, env);
-      return { browserContext, client };
-    });
-  },
-
-  cli: async ({ mcpBrowser }, use, testInfo) => {
-    await use(async (...args: string[]) => {
-      return await runCli(args, { mcpBrowser, testInfo });
-    });
-
-    // Cleanup sessions
-    await runCli(['close-all'], { mcpBrowser, testInfo }).catch(() => {});
-
-    const daemonDir = path.join(testInfo.outputDir, 'daemon');
-    await fs.rm(daemonDir, { recursive: true, force: true }).catch(() => {});
-  },
-});
-
-function cliEnv() {
-  return {
-    PLAYWRIGHT_SERVER_REGISTRY: test.info().outputPath('registry'),
-    PLAYWRIGHT_DAEMON_SESSION_DIR: test.info().outputPath('daemon'),
-    PLAYWRIGHT_SOCKETS_DIR: path.join(test.info().project.outputDir, 'ds', String(test.info().parallelIndex)),
-  };
-}
-
-async function runCli(
-  args: string[],
-  options: { mcpBrowser?: string, testInfo: any },
-): Promise<CliResult> {
-  const stepTitle = `cli ${args.join(' ')}`;
-
-  return await test.step(stepTitle, async () => {
-    const testInfo = options.testInfo;
-
-    // Path to the terminal CLI
-    const cliPath = path.join(__dirname, '../../../node_modules/playwright-core/lib/tools/cli-client/cli.js');
-
-    return new Promise<CliResult>((resolve, reject) => {
-      let stdout = '';
-      let stderr = '';
-
-      const childProcess = spawn(process.execPath, [cliPath, ...args], {
-        cwd: testInfo.outputPath(),
-        env: {
-          ...process.env,
-          ...cliEnv(),
-          PLAYWRIGHT_MCP_BROWSER: options.mcpBrowser,
-          PLAYWRIGHT_MCP_HEADLESS: 'false',
-        },
-        detached: true,
-      });
-
-      childProcess.stdout?.on('data', (data) => {
-        stdout += data.toString();
-      });
-
-      childProcess.stderr?.on('data', (data) => {
-        if (process.env.PWMCP_DEBUG)
-          process.stderr.write(data);
-        stderr += data.toString();
-      });
-
-      childProcess.on('close', async (code) => {
-        await testInfo.attach(stepTitle, { body: stdout, contentType: 'text/plain' });
-        resolve({
-          output: stdout.trim(),
-          error: stderr.trim(),
-        });
-      });
-
-      childProcess.on('error', reject);
-    });
-  });
-}
-
-async function startWithExtensionFlag(browserWithExtension: BrowserWithExtension, startClient: StartClient, env?: Record<string, string>): Promise<Client> {
-  const { client } = await startClient({
-    args: [`--extension`],
-    env,
-    config: {
-      browser: {
-        userDataDir: browserWithExtension.userDataDir,
-      }
-    },
-  });
-  return client;
-}
-
-const testWithOldExtensionVersion = test.extend({
-  pathToExtension: async ({ pathToExtension }, use, testInfo) => {
-    const manifestPath = path.join(pathToExtension, 'manifest.json');
-    const manifest = JSON.parse(await fs.readFile(manifestPath, 'utf8'));
-    manifest.key = extensionPublicKey;
-    manifest.version = '0.0.1';
-    await fs.writeFile(manifestPath, JSON.stringify(manifest, null, 2));
-    await use(pathToExtension);
-  },
-});
+import { test, testWithOldExtensionVersion, expect, extensionId, startWithExtensionFlag } from './extension-fixtures';
 
 test(`navigate with extension`, async ({ startExtensionClient, server }) => {
   const { browserContext, client } = await startExtensionClient();
